Conflicts:
	kingdom/models.py

+++ kingdom/models.py
@@ -43,12 +43,12 @@ class Folk(NamedModel):
 	birth = models.DateTimeField(auto_now_add=True)
 	death = models.DateTimeField(blank=True, null=True)
 
-	fight = models.PositiveSmallIntegerField(validators=[MaxValueValidator(20)], default=0)
-	diplomacy = models.PositiveSmallIntegerField(validators=[MaxValueValidator(20)], default=0)
-	plot = models.PositiveSmallIntegerField(validators=[MaxValueValidator(20)], default=0)
-	scholarship = models.PositiveSmallIntegerField(validators=[MaxValueValidator(20)], default=0)
+	fight = models.PositiveIntegerField(validators=[MaxValueValidator(20)], default=0)
+	diplomacy = models.PositiveIntegerField(validators=[MaxValueValidator(20)], default=0)
+	plot = models.PositiveIntegerField(validators=[MaxValueValidator(20)], default=0)
+	scholarship = models.PositiveIntegerField(validators=[MaxValueValidator(20)], default=0)
 
-	loyalty = models.PositiveSmallIntegerField(validators=[MaxValueValidator(100)], default=0)
+	loyalty = models.PositiveIntegerField(validators=[MaxValueValidator(100)], default=0)
 
 	quality_set = models.ManyToManyField('Quality', blank=True, null=True)
